requirements.txt
# Core LLM and AI Libraries
openai>=1.0.0
anthropic>=0.7.0
huggingface-hub>=0.19.0
transformers>=4.35.0
torch>=2.0.0

# Web Scraping and Processing
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0
selenium>=4.15.0

# Web Development and UI
streamlit>=1.28.0
flask>=3.0.0
fastapi>=0.104.0
uvicorn>=0.24.0

# Data Processing and Analysis
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Jupyter and Interactive Development
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=7.0.0

# Utilities and Helpers
python-dotenv>=1.0.0
tqdm>=4.66.0
click>=8.1.0
rich>=13.7.0

# API and HTTP
httpx>=0.25.0
aiohttp>=3.9.0

# Text Processing
nltk>=3.8.0
spacy>=3.7.0
textstat>=0.7.0

# Configuration and Environment
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Development Tools
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0

# Optional: GPU Support (uncomment if needed)
# cupy-cuda11x>=12.0.0  # For CUDA 11.x
# cupy-cuda12x>=12.0.0  # For CUDA 12.x 